backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "magi-trade"
version = "0.1.0"
description = "Hyperliquid AI-Driven Multi-Agent Trading System"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    # Database
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "mysqlclient>=2.2.0",
    "aiomysql>=0.2.0",
    # Data processing
    "pandas>=2.2.0",
    "pandas-ta>=0.3.14b",
    "numpy>=1.26.0",
    # Hyperliquid SDK
    "hyperliquid-python-sdk>=0.8.0",
    # Scheduler
    "apscheduler>=3.10.0",
    # HTTP client
    "httpx>=0.27.0",
    # Pydantic
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    # AI / LLM
    "langchain>=0.1.0",
    "langchain-anthropic>=0.1.0",
    "langchain-openai>=0.1.0",
    "langgraph>=0.1.0",
    "anthropic>=0.20.0",
    "openai>=1.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.3.0",
    "httpx>=0.27.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["app"]
